Dispatch character creation on the calling being picked
CreateCharacter becomes a dispatcher reading state.factionSlug, so the
page reskins live as a calling is picked and returns to Default when it
is cleared. The Default archetype is the na kit: one responsive
component folding in the phone column that was
mobileArchetypes/DefaultCreateCharacter, which is deleted rather than
left as a second mobile path.

Inks move to the --faction-default-* family because an archetypes/
directory is a faction-dispatched surface, which shrinks the
faction-ink ratchet by one file.

Refs #2346
